Allow inline #CHECK directives and ignore double-commented directives
It can be convenient to comment out an entire block in a test file.  However
CHECK lines were still parsed, so change that. I don't think anyone is using
double commented directives for a good reason.

While we're at it, support inline directives. I'm not sure if FileCheck
supports them but it seems like a good idea for fish.