Print multiple "error annotation lines"
When there is an error on stdout *and* on stderr, we currently only
print one "error annotation line".

That's not great if e.g. there's a large error message.

So instead we just print *all* the error lines. If it turns out to be
a problem, we can add a limit or log file or something.