Create a diff between the lines and the checks
This creates output like

```
Failure in test/files/python_missing_output.py:

  The CHECK on line 8 wants:
    NotPrinted

  which failed to match line stdout:2:
    Second

  Context:
    First
     <= nothing to match CHECK 'NotPrinted' on line 8
    Second
    Third
    1 <= no check matches this, previous check on line 13
    2
    3
    4
    [...] from line 34 (stdout:14):
    6
    6
    6
    Fourth <= does not match CHECK '{{\\d}}' on line 40

  when running command:
    /usr/bin/python test/files/python_missing_output.py
```

(with color).

Instead of just listing the first error (CHECK on line 8 doesn't have a match),
we manage to list the other errors as well, and we also figure out that there's a line *missing*,
instead of chalking it up to "Second" not matching "NotPrinted", or "no more CHECKS left".

Really the breakthrough here is that we go through the entire
list, and if a regex matches *add the literal text* to the list. That
way we can use diffing libs because now it's all just strings.