Hack SequenceMatcher to do the regex matching
This solves an edgecase we actually hit in our tests:

python_missing_output.py complained:

    1 <= no check matches this, previous check on line 13

even tho a CHECK totally *did*, and

    Fourth <= does not match CHECK '{{\\d}}' on line 40

even tho that CHECK already matched the previous number.

The underlying problem was that we'd pass non-matching CHECKs as their
regex-source, which means they could never match the *next* line (or
any other line).

This means that the SequenceMatcher has to know about regex.
Unfortunately, it can't.

What we do is to add an __eq__() method to both CheckCmd and Line that
knows about regex matching, and to force it to be used we override the
__hash__ so it always collides.

Then we pass the actual Line/CheckCmd objects instead of the escaped
test.

This is because python dictionaries first check __hash__, and on
collision check __eq__.

Any better solution would require us to include our own diffing
implementation, as python's difflib unfortunately doesn't have any
better facilities for including this.