Loosen equality checks
This only compares by text, not e.g. line number.

This fixes the issues with regexes being mismatched - apparently
SequenceMatcher is expecting transitivity here in some fashion.

Note that we now hit a bit of a worst case in the test, where it gives
an insert and a delete instead of a replace.

Instead of matching like

```
0 0
a <= not matching '0'
0 0
0 0
```

it now matches like

```
0 0
a <= no CHECK
0 0
0 0
<= no line for '0'
```

Which isn't *wrong* as such, it's just not the ideal change sequence.

This is still preferable to hacking around equals because that would
give arbitrary long false replaces (that we turn into fake matches),
disregarding the context limit.

E.g. if you added more "0" lines and CHECKs it would give

```
0 0
a <= not matching '0'
0 0
0 0
0 0
0 0
0 0
0 0
0 0
0 0
0 0
0 0
0 0
0 0
```

and so on...

This failure mode occurs when you have multiple CHECKs for the same
thing happening in succession, and it will always give a missing line
after the last match.

(yeah, maybe we should do the diffing ourselves)