Skip tests that return 125
Just like `git bisect run` does.

This allows making tests that don't run in all cases.

Specifically what I have in mind is a test to fish that tests `git`,
but that can only work if `git` is installed, and we don't want to
fail it if it doesn't.

The alternative is to fake *all* of the output, which requires
duplicating all of the output somewhere.