Move around output context to stop escaping quotes
Our output used to look like

```
mismatch <= does not match CHECK 'match' on line 12
```

(with color, the "match" in blue, quotes excluded)

Because of this, we needed to escape double/single-quotes or else you
could confuse them with the outer single-quotes.

Instead, we switch it around to

```
mismatch <= does not match CHECK on line 12: match
```

Which doesn't need quotes. We ignore leading/trailing whitespace and
escape embedded newlines already.